requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
pymupdf==1.20.2
python-docx==0.8.11
pillow==10.1.0

# OCR
paddleocr==2.7.0.3
easyocr==1.7.1

# Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LangChain & LLM
langchain==0.1.0
transformers==4.35.2
torch==2.1.0
accelerate==0.24.1

# Utilities
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
opencv-python==4.6.0.66
reportlab==4.0.7
python-dotenv==1.0.0
pydantic==2.5.0

# Monitoring & Logging
colorlog==6.8.0
